requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
plotly>=5.18.0
numpy>=1.26.0
supabase>=2.3.0
python-magic>=0.4.27
PyPDF2>=3.0.0
sendgrid>=6.11.0 